Split out SW mock APIs into their own files
