--- infra/testing/mock-fetch.js
@@ -1,8 +1,0 @@
-module.exports = async (request) => {
-  const response = new Response('Response from mock-fetch.js', {
-    status: 200,
-    statusText: 'ok.',
-  });
-  response.url = request.url;
-  return response;
-};
