+++ infra/testing/sw-env-mocks/Blob.js
@@ -1,0 +1,55 @@
+/*
+ Copyright 2017 Google Inc. All Rights Reserved.
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+
+     http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+*/
+
+// Blob
+// https://w3c.github.io/FileAPI/#dom-blob-blob
+class Blob {
+  constructor(blobParts, options = {}) {
+    if (typeof blobParts === 'undefined') {
+      blobParts = [];
+    }
+    if (!Array.isArray(blobParts)) {
+      throw new TypeError(`Failed to construct 'Blob': ` +
+          `The provided value cannot be converted to a sequence.`);
+    }
+
+    this._parts = blobParts;
+    this._type = options.type || '';
+  }
+
+  get size() {
+    let size = 0;
+    for (const part of this._parts) {
+      size += part.length || part.size;
+    }
+    return size;
+  }
+
+  get type() {
+    return this._type;
+  }
+
+  // Warning: non-standard, but used in other mocks for simplicity.
+  // TODO(philipwalton): implement/use FileReader to get the Blob text.
+  get _text() {
+    let text = '';
+    for (const part of this._parts) {
+      text += typeof part === 'string' ? part : part._text;
+    }
+    return text;
+  }
+}
+
+module.exports = Blob;
