+++ infra/testing/sw-env-mocks/ExtendableEvent.js
@@ -1,0 +1,30 @@
+/*
+ Copyright 2017 Google Inc. All Rights Reserved.
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+
+     http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+*/
+
+
+const Event = require('./Event');
+
+
+// ExtendibleEvent
+// https://www.w3.org/TR/service-workers-1/#extendable-event
+class ExtendableEvent extends Event {
+  waitUntil(/* promise */) {
+    // TODO(philipwalton): keep track of the promises added to each event
+    // and expose them in some way so tests can assert logic doesn't run
+    // after all ExtendableEvents have settled.
+  }
+}
+
+module.exports = ExtendableEvent;
