+++ infra/testing/sw-env-mocks/SyncEvent.js
@@ -1,0 +1,34 @@
+/*
+ Copyright 2017 Google Inc. All Rights Reserved.
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+
+     http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+*/
+
+
+const ExtendableEvent = require('./ExtendableEvent');
+
+// SyncEvent
+// https://wicg.github.io/BackgroundSync/spec/#sync-event
+class SyncEvent extends ExtendableEvent {
+  constructor(type, init = {}) {
+    super(type, init);
+
+    if (!init.tag) {
+      throw new TypeError(
+          `Failed to construct 'SyncEvent': required member tag is undefined.`);
+    }
+
+    this.tag = init.tag;
+    this.lastChance = init.lastChance || false;
+  }
+}
+module.exports = SyncEvent;
