+++ infra/testing/sw-env-mocks/SyncManager.js
@@ -1,0 +1,32 @@
+/*
+ Copyright 2017 Google Inc. All Rights Reserved.
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+
+     http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+*/
+
+
+// SyncManager
+// https://wicg.github.io/BackgroundSync/spec/#sync-manager-interface
+class SyncManager {
+  constructor() {
+    this._tagList = new Set();
+  }
+  async register(tagName) {
+    this._tagList.add(tagName);
+    return Promise.resolve();
+  }
+  async getTags() {
+    return Promise.resolve([...this._tagList]);
+  }
+}
+
+module.exports = SyncManager;
